requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
coverage==6.2
crayons==0.4.0
cryptography==36.0.1
cycler==0.11.0
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.28.3
h11==0.12.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mongomock==3.23.0
mypy-extensions==0.4.3
numpy==1.21.4
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pymongo==4.0.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
selenium==4.1.0
sentinels==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
tomli==1.2.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
webdriver-manager==3.5.2
wheel==0.37.0
Werkzeug==2.0.2
wsproto==1.0.0
